+++ ParqDaoLayer/src/main/java/com/parq/server/dao/ParkingStatusDao.java
@@ -74,23 +74,23 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 	 * @param spaceIds
 	 * @return
 	 */
-	public List<ParkingInstance> getParkingStatusBySpaceIds(int[] spaceIds) {
+	public List<ParkingInstance> getParkingStatusBySpaceIds(long[] spaceIds) {
 
 		if (spaceIds == null || spaceIds.length == 0) {
 			return null;
 		}
 		
 		// create the list of spaces id to check with the cache first
-		List<Integer> spaceIdsToCheck = new ArrayList<Integer>();
-		for (int s : spaceIds) {
+		List<Long> spaceIdsToCheck = new ArrayList<Long>();
+		for (long s : spaceIds) {
 			spaceIdsToCheck.add(s);
 		}
 		
 		
 		List<ParkingInstance> results = new ArrayList<ParkingInstance>();
-		List<Integer> spaceIdsToCheckInDB = new ArrayList<Integer>();
+		List<Long> spaceIdsToCheckInDB = new ArrayList<Long>();
 		// check the cache for the spaces, space by space.
-		for (int spaceId : spaceIdsToCheck)
+		for (long spaceId : spaceIdsToCheck)
 		{
 			ParkingInstance parkInst = getCachedParkingInstanceBySpaceId(spaceId);
 			if (parkInst != null) {
@@ -146,11 +146,11 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 	}
 	
 	
-	private String createCacheKey(String cacheKey, int id) {
+	private String createCacheKey(String cacheKey, long id) {
 		return cacheKey + id;
 	}
 
-	public ParkingInstance getUserParkingStatus(int userId) {
+	public ParkingInstance getUserParkingStatus(long userId) {
 		ParkingInstance userParkingStatus = getCacheParkingInstanceByUserId(userId);
 
 		// if cache match is found, return result
@@ -164,7 +164,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		try {
 			con = getConnection();
 			pstmt = con.prepareStatement(sqlGetParkingStatusByUserId);
-			pstmt.setInt(1, userId);
+			pstmt.setLong(1, userId);
 			ResultSet rs = pstmt.executeQuery();
 
 			List<ParkingInstance> piList = createParkingStatusObject(rs);
@@ -244,7 +244,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 	 */
 	public boolean addNewParkingAndPayment(ParkingInstance parkingInst) {
 		
-		int parkingSpaceId = parkingInst.getSpaceId();
+		long parkingSpaceId = parkingInst.getSpaceId();
 		if (parkingSpaceId < 1) {
 			throw new IllegalStateException("Parking Space Id is invalid: " + parkingSpaceId);
 		}
@@ -259,8 +259,8 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		try {
 			con = getConnection();
 			pstmt = con.prepareStatement(sqlInsertParkingInstance);
-			pstmt.setInt(1, parkingInst.getUserId());
-			pstmt.setInt(2, parkingInst.getSpaceId());
+			pstmt.setLong(1, parkingInst.getUserId());
+			pstmt.setLong(2, parkingInst.getSpaceId());
 			pstmt.setTimestamp(3, new Timestamp(parkingInst.getParkingBeganTime().getTime()));
 			pstmt.setTimestamp(4, new Timestamp(parkingInst.getParkingEndTime().getTime()));
 			pstmt.setBoolean(5, parkingInst.isPaidParking());
@@ -268,7 +268,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 			if (pstmt.executeUpdate() == 1)
 			{
 				pstmt = con.prepareStatement(sqlInsertPayment);
-				pstmt.setInt(1, parkingInst.getSpaceId());
+				pstmt.setLong(1, parkingInst.getSpaceId());
 				pstmt.setString(2, parkingInst.getPaymentInfo().getPaymentType().toString());
 				pstmt.setString(3, parkingInst.getPaymentInfo().getPaymentRefNumber());
 				pstmt.setTimestamp(4, new Timestamp(parkingInst.getPaymentInfo().getPaymentDateTime().getTime()));
@@ -288,7 +288,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		return parkingInstanceCreated;
 	}
 	
-	public boolean refillParkingForParkingSpace(int spaceId, Date newParkingEndTime, Payment payment) {
+	public boolean refillParkingForParkingSpace(long spaceId, Date newParkingEndTime, Payment payment) {
 		
 		if (spaceId < 1 || newParkingEndTime == null) {
 			throw new IllegalStateException(
@@ -297,7 +297,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		}
 		
 		// get the current parkingInstance information for this parking space
-		List<ParkingInstance> curParkInstList = getParkingStatusBySpaceIds(new int[]{spaceId});
+		List<ParkingInstance> curParkInstList = getParkingStatusBySpaceIds(new long[]{spaceId});
 		if (curParkInstList == null || curParkInstList.isEmpty()) {
 			throw new IllegalStateException("Invalid refill require for parking space: " + spaceId);
 		}
@@ -310,8 +310,8 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		try {
 			con = getConnection();
 			pstmt = con.prepareStatement(sqlInsertParkingInstance);
-			pstmt.setInt(1, curParkingInst.getUserId());
-			pstmt.setInt(2, curParkingInst.getSpaceId());
+			pstmt.setLong(1, curParkingInst.getUserId());
+			pstmt.setLong(2, curParkingInst.getSpaceId());
 			pstmt.setTimestamp(3, new Timestamp(System.currentTimeMillis()));
 			pstmt.setTimestamp(4, new Timestamp(newParkingEndTime.getTime()));
 			pstmt.setBoolean(5, curParkingInst.isPaidParking());
@@ -319,7 +319,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 			if (pstmt.executeUpdate() == 1)
 			{
 				pstmt = con.prepareStatement(sqlInsertPayment);
-				pstmt.setInt(1, curParkingInst.getSpaceId());
+				pstmt.setLong(1, curParkingInst.getSpaceId());
 				pstmt.setString(2, payment.getPaymentType().toString());
 				pstmt.setString(3, payment.getPaymentRefNumber());
 				pstmt.setTimestamp(4, new Timestamp(payment.getPaymentDateTime().getTime()));
@@ -339,7 +339,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		return refillComplete;
 	}
 	
-	public boolean unparkBySpaceIdAndParkingInstId(int spaceId, int parkingInstanceId, Date endTime) {
+	public boolean unparkBySpaceIdAndParkingInstId(long spaceId, long parkingInstanceId, Date endTime) {
 		
 		if (spaceId < 1 || parkingInstanceId < 1) {
 			throw new IllegalStateException(
@@ -356,7 +356,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 			con = getConnection();
 			pstmt = con.prepareStatement(sqlUpdateParkingEndTime);
 			pstmt.setTimestamp(1, new Timestamp(endTime.getTime()));
-			pstmt.setInt(2, parkingInstanceId);
+			pstmt.setLong(2, parkingInstanceId);
 			// pstmt.setInt(3, spaceId);
 			if (pstmt.executeUpdate() == 1) {
 				parkingEndTimeUpdated = true;
@@ -377,7 +377,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 	 * Revoke all the cache instance of this User by id, username, and email address.
 	 * @param userID
 	 */
-	private synchronized void revokeSpaceCacheById(int spaceId) {
+	private synchronized void revokeSpaceCacheById(long spaceId) {
 		if (spaceId < 1) {
 			return;
 		}
@@ -389,7 +389,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		}
 
 		// revoke the userId to Parking Instance cache
-		int userId = getCachedSpaceIdToUserId(spaceId);
+		long userId = getCachedSpaceIdToUserId(spaceId);
 		if (userId > 0) {
 			revokeCache(myCache, createCacheKey(getUserParkingStatusCacheKey, userId));
 			// remove the spaceId to userId relationship
@@ -397,7 +397,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		}
 	}
 
-	private ParkingInstance getCachedParkingInstanceBySpaceId(int spaceId) {
+	private ParkingInstance getCachedParkingInstanceBySpaceId(long spaceId) {
 		String cacheKey = createCacheKey(getParkingStatusBySpaceIdsCacheKey, spaceId);
 		ParkingInstance parkInst = null;
 		
@@ -407,7 +407,7 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		return parkInst;
 	}
 	
-	private ParkingInstance getCacheParkingInstanceByUserId(int userId) {
+	private ParkingInstance getCacheParkingInstanceByUserId(long userId) {
 		String cacheKey = createCacheKey(getUserParkingStatusCacheKey, userId);
 
 		ParkingInstance userParkingStatus = null;
@@ -417,12 +417,12 @@ public class ParkingStatusDao extends AbstractParqDaoParent{
 		return userParkingStatus;
 	}
 	
-	private int getCachedSpaceIdToUserId(int spaceId) {
+	private long getCachedSpaceIdToUserId(long spaceId) {
 		String cacheKey = createCacheKey(getSpaceIdByUserId, spaceId);
 		
-		int userId = -1;
+		long userId = -1;
 		if (myCache.get(cacheKey) != null && myCache.get(cacheKey).getValue() != null) {
-			userId = (Integer) myCache.get(cacheKey).getValue();
+			userId = (Long) myCache.get(cacheKey).getValue();
 		}
 		return userId;
 	}
